requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
xgboost>=1.7.0
pyyaml>=6.0
websocket-client>=1.6.0
redis>=4.5.0
pytest>=7.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
